+++ tests/providers/google/cloud/hooks/test_kubernetes_engine.py
@@ -450,3 +450,34 @@ class TestGKEPodHook:
 
     def _get_credentials(self):
         return self.credentials
+
+    @pytest.mark.parametrize(
+        "disable_tcp_keepalive, expected",
+        (
+            (True, False),
+            (None, True),
+            (False, True),
+        ),
+    )
+    @mock.patch(GKE_STRING.format("_enable_tcp_keepalive"))
+    def test_disable_tcp_keepalive(
+        self,
+        mock_enable,
+        disable_tcp_keepalive,
+        expected,
+    ):
+        with mock.patch(
+            BASE_STRING.format("GoogleBaseHook.__init__"), new=mock_base_gcp_hook_default_project_id
+        ):
+            gke_hook = GKEPodHook(
+                gcp_conn_id="test",
+                impersonation_chain=IMPERSONATE_CHAIN,
+                ssl_ca_cert=None,
+                cluster_url=None,
+                disable_tcp_keepalive=disable_tcp_keepalive,
+            )
+        gke_hook.get_credentials = self._get_credentials
+
+        api_conn = gke_hook.get_conn()
+        assert mock_enable.called is expected
+        assert isinstance(api_conn, kubernetes.client.api_client.ApiClient)